requirements.txt
fastapi>=0.120.4
uvicorn[standard]>=0.12.0
sqlalchemy>=2.0.44
pydantic>=2.12.3
passlib[argon2]>=1.7.4
python-jose[cryptography]>=3.5.0
python-multipart>=0.0.20
python-dotenv>=0.13
